+++ src/mainwindow.h
@@ -17,6 +17,7 @@ class AnalysisPanel;
 class MetadataPanel;
 class CompareWorkspace;
 class SearchPanel;
+class BatchDialog;
 class QAction;
 class QMenu;
 class QLineEdit;
@@ -70,6 +71,7 @@ class MainWindow : public QMainWindow
     AnalysisPanel *m_analysisPanel = nullptr;
     MetadataPanel *m_metadataPanel = nullptr;
     SearchPanel *m_searchPanel = nullptr;
+    BatchDialog *m_batchDialog = nullptr;
     CompareWorkspace *m_compareView = nullptr;
 
     QAction *m_actOpenDir = nullptr;
@@ -87,6 +89,7 @@ class MainWindow : public QMainWindow
     QAction *m_actHistoryBack = nullptr;
     QAction *m_actHistoryForward = nullptr;
     QAction *m_actToggleSearch = nullptr;
+    QAction *m_actBatch = nullptr;
     QMenu *m_recentMenu = nullptr;
     QMenu *m_recentFileMenu = nullptr;  // recent-files menu (opened images)
     QMenu *m_favMenu = nullptr;